Fixed the PDF link for publications
For links that contain several dots, such as a DOI, the 'relURL' method returned an incorrect value ('/url' instead of 'url').
This behaviour seems to happen only when deploying ('hugo --minify'), not in dev server ('hugo server').
This commit replaces 'relURL' with 'absURL', which seems to work in both contexts.