Better events (talks) listing
Wowchemy does not recognize "talk" as the correct type. We use instead "event" as the folder's name, which is correctly recognized and renders with the appropriate view.
Added configuration to make sure that the links still remain to "talk/..." instead of "event/...".
Also added a header image to the events' listing.