Add research activities
The previous Research page included only a list of publications.
It lacked the research activities (projects) and mentored students.

This commit adds a new page with a layout to include these activities.
The menu is also changed to redirect to this new page; the page includes a link to the publications page.